Refactor Entities and DTO's with nullables only on the appropiate properties and converting DTO doubles that always ended on .0's into proper ints on the Entity

+++ LifeTracker/Services/GarminBridgeService.cs
@@ -334,8 +334,8 @@ namespace LifeTracker.Services;
                 LightSleepSeconds = d.LightSleepSeconds,
                 RemSleepSeconds = d.RemSleepSeconds,
                 AwakeSleepSeconds = d.AwakeSleepSeconds,
-                AvgHeartRate = d.AvgHeartRate,
-                AvgSleepStress = d.AvgSleepStress
+            AvgHeartRate = (int)d.AvgHeartRate,
+            AvgSleepStress = (int)d.AvgSleepStress,
             };
     }
